api/pyproject.toml
[project]
name = "jet-rag-api"
version = "0.1.0"
description = "Jet-Rag 백엔드 — 한국어 멀티포맷 RAG 개인 지식 에이전트"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.136.0",
    "filetype>=1.2.0",
    "google-genai>=1.0.0",
    "olefile>=0.47",
    "pillow>=12.2.0",
    "pyhwp>=0.1b15",
    "pymupdf>=1.24.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
    "python-hwpx>=2.9.0",
    "python-multipart>=0.0.19",
    "supabase>=2.11.0",
    "trafilatura>=2.0.0",
    "uvicorn[standard]>=0.45.0",
]
